pyproject.toml
[tool.poetry]
name = "probograph"
version = "0.1.0"
description = ""
authors = ["luisferreiraa <luis.c.ferreira.prt_a@msft.cesae.pt>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
sqlmodel = "^0.0.22"
strawberry-graphql = {extras = ["fastapi"], version = "^0.258.0"}
passlib = "^1.7.4"
bcrypt = "^4.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
